# Use wide-word Word128 to implement IPv6 addresses.