--- a/ip.cabal
+++ b/ip.cabal
@@ -55,6 +55,7 @@ library
     , primitive >= 0.6 && < 0.7
     , text >= 1.2  && < 1.3
     , vector >= 0.11 && < 0.13
+    , wide-word >= 0.1.0.8 && < 0.2
   ghc-options: -Wall -O2
   default-language: Haskell2010
 
@@ -65,6 +66,7 @@ test-suite test
   build-depends:
       base
     , ip
+    , wide-word
     , test-framework
     , test-framework-quickcheck2
     , QuickCheck
@@ -89,6 +91,7 @@ test-suite spec
   build-depends:
       base
     , ip
+    , wide-word
     , hspec >= 2.5.5
   other-modules:
     Net.IPv4Spec
@@ -104,6 +107,7 @@ test-suite doctest
   build-depends:
       base
     , ip
+    , wide-word
     , doctest >= 0.10
     , QuickCheck
   default-language:    Haskell2010