@andrewthad 2019-03-23
This looks pretty good. Also, good job paying careful attention to avoid performance degradation. I’m on a phone right now, so I’ll review this a little more thoroughly later.

As a reminder to myself, this will require a major version bump since the data constructor for IPv6 has changed.