@andrewthad 2019-03-23
Actually, just go ahead and bump the version in the PR. And add a note to the changelog.