@andrewthad 2019-03-23
Released as [ip-1.5.0](http://hackage.haskell.org/package/ip-1.5.0).